h265nal: fix bug on SCC extension management
The palette predictor initializer for both sps_scc_extension and
pps_scc_extension was broken. In both cases, the palette has 1 or 3
vectors (one per component), each of which is composed of a fixed amount
of initalizers. The palette predictor initializer must therefore be
indexed based on the comp value.

Not sure how to test this without actual content, and this seems to
be an obscure feature.