h265nal: fix incoherence in pred_weight_table
pred_weight_table() definition in 2018 standard is incoherent. The table
is defined as:

```
pred_weight_table() {
  ...
  if (ChromaArrayType != 0) {
    for (i = 0; i <= num_ref_idx_l0_active_minus1; i++) {
      if (...) {
        chroma_weight_l0_flag[i]  // u(1)
      }
    }
  }

  for (i = 0; i <= num_ref_idx_l0_active_minus1; i++) {
    ...
    if (chroma_weight_l0_flag[i]) {
      for( j = 0; j < 2; j++ ) {
        delta_chroma_weight_l0[i][j]  // se(v)
        delta_chroma_offset_l0[i][j]  // se(v)
      }
    }
  }
  ...
```

I added several curly brackets to make clearer the if/for blocks. Note
that the `chroma_weight_l0_flag[i]` vector values are only defined when
`(ChromaArrayType != 0)`, but the values of the vector are latter used
without checking that `(ChromaArrayType != 0)`.

Solution is to check for `(ChromaArrayType != 0)` before accessing
`chroma_weight_l0_flag`.

Also removed fuzzing crashes.