h265nal: ignore general_profile_compatibility_flag flags
All profile checks include "general_profile_idc equal to j or
general_profile_compatibility_flag[j] equal to 1". Moreover, it also
states that "general_profile_compatibility_flag[ j ] equal to 1, when
general_profile_space is equal to 0, indicates that the CVS conforms
to the profile indicated by general_profile_idc equal to j as specified
in Annex A. When general_profile_space is equal to 0,
general_profile_compatibility_flag[ general_profile_idc ] shall be
equal to 1.".

As a safe approach, we will check for general_profile_space to be 0,
and then ignore the profile_compatibility_flag[j] flags.